+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - replace D1 with KV
 - prevent FOUC on load/refresh
 - fix text contrast issues for accessibility
+- use virtual table
 
 ## [0.4.1] - 2025-10-24